move get element style values helper
