+++ lib/shared/helpers/styles-manager.js
@@ -3,7 +3,7 @@ import find from 'lodash/find';
 import forEach from 'lodash/forEach';
 import {Stylesheet} from 'relax-jss';
 
-import getElementStyleValues from './get-element-style-values';
+import getElementStyleValues from './element/get-style-values';
 
 class StylesManager {
   constructor () {
