+++ lib/shared/screens/admin/shared/components/page-builder-menu/tabs/style/style-picker/editing/index.js
@@ -2,7 +2,7 @@ import * as pageBuilderActions from 'actions/page-builder';
 
 import bind from 'decorators/bind';
 import forEach from 'lodash/forEach';
-import getElementStyleValues from 'helpers/get-element-style-values';
+import getElementStyleValues from 'helpers/element/get-style-values';
 import union from 'lodash/union';
 import Component from 'components/component';
 import React, {PropTypes} from 'react';
